Isolate (some) runtime const functions into a runtime module
