+++ opal/runtime/boot.js
@@ -273,25 +273,6 @@
     }
   }
 
-  function handle_autoload(cref, name) {
-    if (!cref.$$autoload[name].loaded) {
-      cref.$$autoload[name].loaded = true;
-      try {
-        Opal.Kernel.$require(cref.$$autoload[name].path);
-      } catch (e) {
-        cref.$$autoload[name].exception = e;
-        throw e;
-      }
-      cref.$$autoload[name].required = true;
-      if (cref.$$const[name] != null) {
-        cref.$$autoload[name].success = true;
-        return cref.$$const[name];
-      }
-    } else if (cref.$$autoload[name].loaded && !cref.$$autoload[name].required) {
-      if (cref.$$autoload[name].exception) { throw cref.$$autoload[name].exception; }
-    }
-  }
-
   // Constants
   // ---------
   //
@@ -309,28 +290,15 @@
     if (cref) {
       if (cref.$$const[name] != null) { return cref.$$const[name]; }
       if (cref.$$autoload && cref.$$autoload[name]) {
-        return handle_autoload(cref, name);
+        return Opal.handle_autoload(cref, name);
       }
     }
   }
 
-  // Walk up the nesting array looking for the constant
-  function const_lookup_nesting(nesting, name) {
-    var i, ii, constant;
+  Opal.const_get_name = const_get_name;
 
-    if (nesting.length === 0) return;
-
-    // If the nesting is not empty the constant is looked up in its elements
-    // and in order. The ancestors of those elements are ignored.
-    for (i = 0, ii = nesting.length; i < ii; i++) {
-      constant = nesting[i].$$const[name];
-      if (constant != null) {
-        return constant;
-      } else if (nesting[i].$$autoload && nesting[i].$$autoload[name]) {
-        return handle_autoload(nesting[i], name);
-      }
-    }
-  }
+  // Initialize the top level constant cache generation counter
+  Opal.const_cache_version = 1;
 
   // Walk up the ancestors chain looking for the constant
   function const_lookup_ancestors(cref, name) {
@@ -344,105 +312,12 @@
       if (ancestors[i].$$const && $has_own(ancestors[i].$$const, name)) {
         return ancestors[i].$$const[name];
       } else if (ancestors[i].$$autoload && ancestors[i].$$autoload[name]) {
-        return handle_autoload(ancestors[i], name);
+        return Opal.handle_autoload(ancestors[i], name);
       }
     }
   }
 
-  // Walk up Object's ancestors chain looking for the constant,
-  // but only if cref is missing or a module.
-  function const_lookup_Object(cref, name) {
-    if (cref == null || cref.$$is_module) {
-      return const_lookup_ancestors(_Object, name);
-    }
-  }
-
-  // Call const_missing if nothing else worked
-  function const_missing(cref, name) {
-    return (cref || _Object).$const_missing(name);
-  }
-
-  // Look for the constant just in the current cref or call `#const_missing`
-  Opal.const_get_local = function(cref, name, skip_missing) {
-    var result;
-
-    if (cref == null) return;
-
-    if (cref === '::') cref = _Object;
-
-    if (!cref.$$is_module && !cref.$$is_class) {
-      $raise(Opal.TypeError, cref.toString() + " is not a class/module");
-    }
-
-    result = const_get_name(cref, name);
-    return result != null || skip_missing ? result : const_missing(cref, name);
-  };
-
-  // Look for the constant relative to a cref or call `#const_missing` (when the
-  // constant is prefixed by `::`).
-  Opal.const_get_qualified = function(cref, name, skip_missing) {
-    var result, cache, cached, current_version = Opal.const_cache_version;
-
-    if (name == null) {
-      // A shortpath for calls like ::String => $$$("String")
-      result = const_get_name(_Object, cref);
-
-      if (result != null) return result;
-      return Opal.const_get_qualified(_Object, cref, skip_missing);
-    }
-
-    if (cref == null) return;
-
-    if (cref === '::') cref = _Object;
-
-    if (!cref.$$is_module && !cref.$$is_class) {
-      $raise(Opal.TypeError, cref.toString() + " is not a class/module");
-    }
-
-    if ((cache = cref.$$const_cache) == null) {
-      $prop(cref, '$$const_cache', Object.create(null));
-      cache = cref.$$const_cache;
-    }
-    cached = cache[name];
-
-    if (cached == null || cached[0] !== current_version) {
-      ((result = const_get_name(cref, name))              != null) ||
-      ((result = const_lookup_ancestors(cref, name))      != null);
-      cache[name] = [current_version, result];
-    } else {
-      result = cached[1];
-    }
-
-    return result != null || skip_missing ? result : const_missing(cref, name);
-  };
-
-  // Initialize the top level constant cache generation counter
-  Opal.const_cache_version = 1;
-
-  // Look for the constant in the open using the current nesting and the nearest
-  // cref ancestors or call `#const_missing` (when the constant has no :: prefix).
-  Opal.const_get_relative = function(nesting, name, skip_missing) {
-    var cref = nesting[0], result, current_version = Opal.const_cache_version, cache, cached;
-
-    if ((cache = nesting.$$const_cache) == null) {
-      $prop(nesting, '$$const_cache', Object.create(null));
-      cache = nesting.$$const_cache;
-    }
-    cached = cache[name];
-
-    if (cached == null || cached[0] !== current_version) {
-      ((result = const_get_name(cref, name))              != null) ||
-      ((result = const_lookup_nesting(nesting, name))     != null) ||
-      ((result = const_lookup_ancestors(cref, name))      != null) ||
-      ((result = const_lookup_Object(cref, name))         != null);
-
-      cache[name] = [current_version, result];
-    } else {
-      result = cached[1];
-    }
-
-    return result != null || skip_missing ? result : const_missing(cref, name);
-  };
+  Opal.const_lookup_ancestors = const_lookup_ancestors;
 
   // Register the constant on a cref and opportunistically set the name of
   // unnamed classes/modules.
@@ -467,11 +342,11 @@
     // Add a short helper to navigate constants manually.
     // @example
     //   Opal.$$.Regexp.$$.IGNORECASE
-    if (cref === Opal) {
+    if (cref === Opal && cref.$$) {
       // Opal is now the same as the Opal constant.
       // Unfortunately, we already use $$ function on Opal, so this simply
       // sets name on this function for compatibility.
-      cref.$$[name] = value; 
+      cref.$$[name] = value;
     }
     else {
       cref.$$ = cref.$$const;
@@ -493,65 +368,6 @@
   };
 
   Opal.const_set = $const_set;
-
-  // Get all the constants reachable from a given cref, by default will include
-  // inherited constants.
-  Opal.constants = function(cref, inherit) {
-    if (inherit == null) inherit = true;
-
-    var module, modules = [cref], i, ii, constants = {}, constant;
-
-    if (inherit) modules = modules.concat($ancestors(cref));
-    if (inherit && cref.$$is_module) modules = modules.concat([Opal.Object]).concat($ancestors(Opal.Object));
-
-    for (i = 0, ii = modules.length; i < ii; i++) {
-      module = modules[i];
-
-      // Do not show Objects constants unless we're querying Object itself
-      if (cref !== _Object && module == _Object) break;
-
-      for (constant in module.$$const) {
-        constants[constant] = true;
-      }
-      if (module.$$autoload) {
-        for (constant in module.$$autoload) {
-          constants[constant] = true;
-        }
-      }
-    }
-
-    return Object.keys(constants);
-  };
-
-  // Remove a constant from a cref.
-  Opal.const_remove = function(cref, name) {
-    Opal.const_cache_version++;
-
-    if (cref.$$const[name] != null) {
-      var old = cref.$$const[name];
-      delete cref.$$const[name];
-      return old;
-    }
-
-    if (cref.$$autoload && cref.$$autoload[name]) {
-      delete cref.$$autoload[name];
-      return nil;
-    }
-
-    $raise(Opal.NameError, "constant "+cref+"::"+cref.$name()+" not defined");
-  };
-
-  // Generates a function that is a curried const_get_relative.
-  Opal.const_get_relative_factory = function(nesting) {
-    return function(name, skip_missing) {
-      return Opal.$$(nesting, name, skip_missing);
-    }
-  }
-
-  // Setup some shortcuts to reduce compiled size
-  Opal.$$ = Opal.const_get_relative;
-  Opal.$$$ = Opal.const_get_qualified;
-  Opal.$r = Opal.const_get_relative_factory;
 
   function descends_from_bridged_class(klass) {
     if (klass == null) return false;
