+++ opal/runtime/runtime.rb
@@ -1,3 +1,4 @@
 require 'runtime/boot'
+require 'runtime/const'
 require 'runtime/hash'
 require 'runtime/helpers'
